cadastro de produtos

+++ VirtusGo/src/VirtusGo.Core.Application/ViewModels/ProdutoViewModel.cs
@@ -5,7 +5,7 @@ namespace VirtusGo.Core.Application.ViewModels
         public int Id { get; set; }
         public string Descricao { get; set; }
         public string Unidade { get; set; }
-        public int ValorUnitario { get; set; }
+        public double ValorUnitario { get; set; }
         public int Estoque { get; set; }
         public string NCM { get; set; }
     }
