+++ VirtusGo/src/VirtusGo.Core.Domain/Produtos/Commands/RegistrarProdutoCommand.cs
@@ -10,7 +10,7 @@ namespace VirtusGo.Core.Domain.Produtos.Commands
             int id,
             string descricao,
             string unidade,
-            int valorUnitario,
+            double valorUnitario,
             int estoque,
             string ncm
             )
